+++ Code/Player/Enums.hpp
@@ -3,20 +3,20 @@
 #include <cstdint>
 
 // See <a href="http://wiki.sa-mp.com/wiki/OnPlayerDisconnect">OnPlayerDisconnect on SA-MP Wiki</a>.
-enum class DisconnectReason : uint16_t
+enum class DisconnectReason : uint8_t
 {
 	kTimeout = 0, // The player's connection was lost. Either their game crashed or their network had a fault. 
 	kQuit, // The player purposefully quit, either using the /quit (/q) command or via the pause menu.
 	kKick // The player was kicked or banned by the server. 
 };
 
-enum class PlayerSex : uint16_t
+enum class PlayerSex : uint8_t
 {
 	kMale = 1,
 	kFemale
 };
 
-enum class PlayerData : uint16_t
+enum class PlayerData : uint8_t
 {
 	// Variables from MySQL.
